+++ ts/features/bonus/siciliaVola/screens/voucherGeneration/WorkerSelectDestinationScreen.tsx
@@ -16,7 +16,7 @@ import {
   svGenerateVoucherSelectCompany
 } from "../../store/actions/voucherGeneration";
 import FooterWithButtons from "../../../../../components/ui/FooterWithButtons";
-import { selectedBeneficiaryCategorySelector } from "../../store/reducers/voucherRequest";
+import { selectedBeneficiaryCategorySelector } from "../../store/reducers/voucherGeneration/voucherRequest";
 import { Company } from "../../types/SvVoucherRequest";
 import { navigateToSvSelectFlightsDateScreen } from "../../navigation/actions";
 import I18n from "../../../../../i18n";
